# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Finca Farm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List farms for current user
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Create a new farm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmRequest"
      responses:
        "201":
          description: Created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{farmId}:
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get farm details
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    put:
      operationId: updateFarm
      tags: [farms]
      summary: Update farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFarmRequest"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{farmId}/members:
    get:
      operationId: listFarmMembers
      tags: [farms]
      summary: List farm members
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FarmMember"
    post:
      operationId: inviteFarmMember
      tags: [farms]
      summary: Invite a member to farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmMember"
  /farms/{farmId}/members/{userId}:
    delete:
      operationId: removeFarmMember
      tags: [farms]
      summary: Remove a member from farm
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed
  /farms/{farmId}/stats:
    get:
      operationId: getFarmStats
      tags: [farms]
      summary: Get farm dashboard stats
      parameters:
        - name: farmId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Farm stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    Profile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [owner, worker]
        preferredLanguage:
          type: string
        createdAt:
          type: string
      required:
        - id
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        preferredLanguage:
          type: string
    Farm:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        location:
          type: string
        totalHectares:
          type: number
        createdAt:
          type: string
        userRole:
          type: string
      required:
        - id
        - name
    CreateFarmRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        totalHectares:
          type: number
      required:
        - name
    UpdateFarmRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        totalHectares:
          type: number
    FarmMember:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        userId:
          type: string
        role:
          type: string
        permissions:
          type: object
        profile:
          $ref: "#/components/schemas/Profile"
        createdAt:
          type: string
      required:
        - id
        - farmId
        - userId
    InviteMemberRequest:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [owner, worker]
      required:
        - email
    FarmStats:
      type: object
      properties:
        totalAnimals:
          type: integer
        animalsBySpecies:
          type: array
          items:
            type: object
            properties:
              species:
                type: string
              count:
                type: integer
        lowStockCount:
          type: integer
        upcomingMedicalCount:
          type: integer
        recentActivityCount:
          type: integer
        upcomingMedical:
          type: array
          items:
            $ref: "#/components/schemas/MedicalRecord"
        lowStockItems:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
      required:
        - totalAnimals
        - lowStockCount
...
    MedicalRecord:
      type: object
      properties:
        id:
          type: string
        animalId:
          type: string
        recordType:
          type: string
          enum: [vaccination, treatment, checkup, surgery, deworming, other]
        title:
          type: string
        description:
          type: string
        vetName:
          type: string
        costCop:
          type: number
        recordDate:
          type: string
        nextDueDate:
          type: string
        createdAt:
          type: string
        animalName:
          type: string
        animalTag:
          type: string
      required:
        - id
        - animalId
        - recordType
        - title
        - recordDate
...
    InventoryItem:
      type: object
      properties:
        id:
          type: string
        farmId:
          type: string
        category:
          type: string
          enum: [feed, medicine, tools, supplies]
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
          enum: [kg, liters, units, bags, doses, other]
        lowStockThreshold:
          type: number
        costPerUnitCop:
          type: number
        supplierName:
          type: string
        supplierContact:
          type: string
        expirationDate:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [ok, low, expired]
        createdAt:
          type: string
      required:
        - id
        - farmId
        - category
        - name
        - quantity
        - unit
...