# lib/api-zod/src/generated/types/farmMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finca Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { FarmMemberPermissions } from "./farmMemberPermissions";
import type { Profile } from "./profile";

export interface FarmMember {
  id: string;
  farmId: string;
  userId: string;
  role?: string;
  permissions?: FarmMemberPermissions;
  profile?: Profile;
  createdAt?: string;
}